fix: stop .js output appearing in dist
ts-loader is quirky. If you use some outFiles in a bundle, they don't
end up in dist, but if you use them indirectly (via a project reference)
then they DO end up in dist.

The simple solution is to split the output with the declaration files
going directly to dist and the output going to build.