feat: add test runner (#317)
* feat: switch to project references

This had the side effect of moving the declartion files from being
siblings of the bundle to being inside /lib. This should be fine, but it
is a change.

* feat: copy over test-runner code

* fix: add exports property for new and old bundles

* fix(ci): install chrome for puppeteer

* fix: export correct files

* fix: use --build for type checking

Otherwise the lint script will only check the base project, not any of
the referenced ones.

* fix: stop .js output appearing in dist

ts-loader is quirky. If you use some outFiles in a bundle, they don't
end up in dist, but if you use them indirectly (via a project reference)
then they DO end up in dist.

The simple solution is to split the output with the declaration files
going directly to dist and the output going to build.

* fix: ensure that all declaration files are outputted

* fix(tests): ignore build folders when testing

* refactor: use workspaces to allow for multiple React versions

Also, make it explicit which React version is used in the DOM evaluator.

* test: cover more of Enzyme's api

Turns out shallow works (at least somewhat) with React 18, but mount
does not.

* feat: add more exports

The test-runner export was wrong and the client wants to consume
package.json so that needs exporting.

* refactor: rename FCCSandbox -> FCCTestRunner

* chore: jest -> vitest

Why? Because it is hard to get Jest to work with a TS project that has
both project references and workspaces. Vitest just seems to work.

* fix: finish migrating to vitest

* fix: stop using deprecated vitest config

* fix: flaky test

Vitest is so fast that it can run the test before the audio element has
finished loading and thus generate a different error.

* fix: teardown puppeteer after testing