feat: use function constructor (JS + DOM)
The function constructor modularize the code a little, since it's now
possible to create a scope for the evaluation separately from the eval
call. The previous implementation had to declare variables in the
surrounding scope.

There is a breaking change because the function returns nothing. This is
in contrast to eval with returned the last call. As a result all the
tests that were "returning" (i.e. had as their last call) a function,
will need to inline that function.