fix: use concurrently to give correct watch behaviour
Vitest has projects which are a great way to organise test config, but
watchTriggerPatterns is global, not per project. If you want a single
project to behave like e2e tests (i.e. run if the source changes)
and another like unit tests (i.e. run if a file they're importing
changes), it doesn't work.

Concurrently fixes this because it handles running the projects and they
handle (separately) which tests should run.